backend/requirements.txt
neomodel~=5.1.2
django_neomodel
djangorestframework~=3.14.0
Django~=5.0.1